Feature/NQ-933: First pass register abstraction layer (#8)
Provides a minimal implementation of a Register Abstraction Layer which follows (some of) the UVM  principles. This change introduces a `RegModel` class which provides methods for reading and writing to registers/fields defined within a SystemRDL spec. To permit driving a simulation DUT with this class it accepts a set of read/write callbacks which perform the reading/writing of a register bus in a domain specific way (i.e. a coco.tb driver).