bugfix(templates): fix for uppercase names (#15)
* bugfix(templates): fix for uppercase names

This commit fixes an issue in the Mako templates where the
register/field names were used as is.  When using uppercase names, we
ended up with a mix of lowercase and uppercase strings used in the RTL.

This commit also cleans up the old Migen based CSR interface that was
put in as a temporary hack during development.  The software interface
is now a simple read/write enable pair with address and read/write data.
There is no support for a write mask and any bridging to common bus
protocols is assumed to be done in the instantiating level of hierarchy.

* fixup: fixing output logic type.

* fixup: updated testbench with new signal names